requirements.txt
flask
pillow
torch
git+https://github.com/xinntao/Real-ESRGAN.git
